+++ include/libfly/system/property.hpp
@@ -222,6 +222,18 @@ namespace fly {
     };
 
     /**
+     * @brief Tag type for a particle local environment hash.
+     */
+    struct Hash : system::Property<std::size_t> {
+      static constexpr char const* tag = "log/particles/hash";  ///< GSD chunk label.
+    };
+
+    /**
+     * @brief Type literal.
+     */
+    inline constexpr Hash hash_;
+
+    /**
      * @brief Type literal.
      */
     inline constexpr Type tp_;